extracted the load utility logic where it should be, also using the '_utils' name so that no one has the idea to use it in a test folder
