+++ h2o-py/tests/__init__.py
@@ -5,5 +5,6 @@ from __future__ import absolute_import
 from .pyunit_utils import *  # NOQA
 from .pydemo_utils import *  # NOQA
 from .pybooklet_utils import *  # NOQA
+from ._utils import *
 
 load_utilities()
